Clean up interface network assignment
The networktype field is no longer used in the interface table
since an interface can have multiple networks associated with it.
This commit is to remove the networktype field and use the
network association instead.

The host interface commands used to allow associating an initial
network to the interface. This has been changed such that network
association can only be done with the interface-network commands.

The address assignment of an interface is performed when a network
is assigned to an interface, not when the interface is provisioned.

Data interface no longer requires at least one datanetwork
associated with it.

The REST API doc is also updated to reflect the changes.

Story: 2004273
Task: 30947
Task: 33504

Change-Id: I952008408826f4c630246477d71687628357622c
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>