Use derived docker_no_proxy values if not provided
Currently, if the user does not specify the docker_no_proxy list in
the override file, NO_PROXY parameter is empty in http-proxy.conf
file.

In this commit, the NO_PROXY parameter is set using the combined
docker_no_proxy list, which is a merged list of values derived
from network config and user provided values if there are any.

Closes-Bug: 1830436
Change-Id: Ia72ce95bfbf5d06e14e8d2aa18c048f3c344b768
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>