playbookconfig: add variable include for site overrides
Some playbook variables are site specific rather than host specific. For
example "dns_servers" versus "external_oam_unit_0_address". It is
cumbersome for an end user to have to override site specific variable in
N different host files when it applies to all hosts in a specific site.

The main play of the playbook is being extended to add an additional
include so that a "site.yml" file can be imported with variable
overrides specific to an entire site.

The new "include_vars" is performed ahead of the host level
"include_vars" so that if an end user still wants to override a site
specific variable on a per-host basis that is still possible.

Change-Id: I0ac49e9311af24d88d29dc381168a60c38f642cf
Closes-Bug: #1832045
Signed-off-by: Allain Legacy <allain.legacy@windriver.com>