Relocate Ansible config and inventory files
In this commit, the Ansible config and default inventory
files are moved out of bootstrap playbook to the parent
directory so they can be shared by all StarlingX playbooks.

In addition, the default inventory file has the following
updates:
  - hosts definition is changed to "all" from "bootstrap"
  - ansible_become definition is removed as this setting
    would affect ALL hosts including Ansible control host
    and thus break remote play of the bootstrap playbook.
  - password_change_responses definition is moved to
    bootstrap default.yml as it is specific to bootstrap.

Bootstrap playbook is updated to adapt to files relocation
and inventory file updates.

Tests:
  - Local play, replay, remote play of bootstrap playbook
    with and without secret file.

Story: 2004695
Task: 33567
Change-Id: I8aa4e554904f7c1254447d2bdd9e7dd484b5f216
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>