Scope error handling in override files import
Previously, the import of variables in overrides files was done in
best effort fashion. If the override files did not exist, bootstrap
defaults would be applied which is a desirable behaviour. However, if
there was a syntax error in one (or more) of the override files, the
affected file(s) would be skipped. This is an undesirable behavior.

This commit fails the bootstrap if any vars import failed due to syntax/
formatting errors.

Closes-Bug: 1832733
Depends-On: I8aa4e554904f7c1254447d2bdd9e7dd484b5f216
Change-Id: I183d3175e5bbdc470f82068eaed868d70921d326
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>