B&R: Wipe ceph OSDs from Ansible bootstrap playbook
To keep ceph-mon data intact at restore, ceph osds should not be wiped.
Every ceph osds and journals have a Globally unique identifier(GUID)
and based on this GUID we wipe partitions and data at the beginning
and at the end.

In ansible playbook we add a configuration option that by default
is false:
 o 'wipe_ceph_osds: false' means that ceph data is recovered,
   ceph osds are not wiped.
 o 'wipe_ceph_osds: true' means that ceph data is not recovered,
   ceph osds are wiped.

Ceph osds should be wiped on a fresh install or a full system restore
and not wiped on a partial system restore.
In order to wipe the ceph osds on a partial system restore the
restore playbook should be issued with 'wipe_ceph_osds=true':
"ansible-playbook /usr/share/ansible/stx-ansible/playbooks/bootstrap/restore_platform.yml
-e "wipe_ceph_osds=true""

Co-authored-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>
Change-Id: Id56789db11c1fb180608975a962baf19514d8da6
Story: 2004761
Task: 34284
Depends-On: Ief93da56c1344ab259e674ca8d79f9952fefcd2b
Signed-off-by: Elena Taivan <elena.taivan@windriver.com>
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>