Add customer-specified certificates for kubernetes
We need the ability to update the Kubernetes ApiServer RootCA at
ansible-bootstrap-time. This includes the ability of being able to
specify the apiServerCertSANs such that user can specify additional
DNS:<FQDN> and/or IP Records for the auto-generated
apiServerCertificate.

This change adds support for specifying the kubernetes certificates and
apiserver certificate SANs in ansible.

Depends-On: https://review.opendev.org/#/c/671559/
Closes-Bug: 1837079
Change-Id: I5528c8ddcdd3203dfdae8c63a944957d424e4158
Signed-off-by: David Sullivan <david.sullivan@windriver.com>