Remove osd from the restored ceph crushmap
When loading the ceph crushmap during platform restore, ceph_mon
crashes. This is due to upstream behavior where it is expected that
if you load a crushmap with OSDs defined they are actually reflected
in the osdmap. If they are not ceph_mon will crash when 'ceph osd tree'
is called by ceph-manager when checking ceph cluster health. The solution
is to remove all the OSDs from the restored crushmap. When OSDs are
created either during host unlock or via runtime manifest, they will be
re-inserted into the crushmap.

Change-Id: I5b98706b1e561f028d72f1f8028a84ce75851ef4
Story: 2004761
Task: 35965
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>