Revert "Revert "Changing tiller pod networking settings to improve swact time""
This reverts commit fe10dcbfed9fd4a6b0e4494cd6d414bf78f03bab.

It was thought that setting the TCP timeouts for the cluster
network was enough to address the issues with the helm commands
hanging after a controller swact. This is not the case. In
particular, swacting away from the controller with the
tiller-deploy pod seems to cause tcp connection from that pod to
the kube-apiserver to hang. Putting the tiller-deploy pod back on
the host network "fixes" the issue.

Change-Id: I89c4db6dc063f238c70fad4e913577046e5452f5
Related-Bug: 1817941
Partial-Bug: 1837055
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>