Configure multus with the portmap plugin
With the introduction of Multus CNI, the portmap chained plugin
was no longer configured.

This commit simply adds the portmap plugin to the Multus configuration
template.

The portmap plugin allows a user to create a pod/deployment with a
host->container port mapping by including the hostPort, containerPort
in the container specification.  ie:

spec:
  containers:
    ports:
    - containerPort: 80
      hostPort: 8089

In this example, one could access the service in the pod via
container_ip:80 or node_ip:8089

Closes-Bug: 1832892

Change-Id: I0549bff242df881e652b742abdf4fb342b5d7a5e
Signed-off-by: Steven Webster <steven.webster@windriver.com>