ansible replay fails if kubeadm init was not successful
This is a cherry-pick of:
https://review.opendev.org/#/c/678649/
with the Change Id and Bug Id updated.

k8s 1.16 re-introduced this bug.

Also, a kube reset triggers the unmount before it does its rm,
which means the rm in ansible is a requirement.

The cherry pick had to be changed because the rm is required,
and the 'when' clause is applied per task in the block
rather than the entire block.

Original commit message text:

Ansible replay is broken since the introduction of kubelet-fs
https://opendev.org/starlingx/config/commit/e74ef5f7c4c71464347142932615e1545884133c

The kubelet-fs change uncovered a flaw in kubeadm reset
https://github.com/kubernetes/kubeadm/issues/1294

kubeadm will sometimes unmount /var/lib/kubelet/. To correct this we
will remount the kubelet-lv on ansible replays and wipe the contents.

Change-Id: Ief5990784e6af5ac01612552aa7a11b1493ac7f6
Closes-Bug: 1847147
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>