Multus support for IPv6 service endpoint
The K8s service host in the Multus kubeconfig file is currently
not wrapped with [brackets] in the case an IPv6 cluster service
endpoint has been configured.

This causes issues for Multus when it attemps to get (curl) for
the address.

This fix ensures the IPv6 address is properly formatted for use
by Multus.

Closes-Bug: 1836972

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I7476908895d22eaa117a6eef85eff4c641a43d31