Enable default CNI plugins
Currently, only a subset of the default CNI plugins are enabled in
StarlingX.  These plugins are provided by Calico and placed in the
/opt/cni/bin directory.

On CentOS, the full set of CNI plugins that are distributed by the
containernetworking-plugins package are installed under the
default directory for system daemons and utilities at
/usr/libexec/cni

This means that some of the default plugins such as static ipam and
dhcp are not available to be used, as kublet looks at the
/opt/cni/bin directory by default.

This commit changes the cni-bin volume of the Calico, Multus, and
SRIOV CNI daemonsets to refer to the plugin executable path used by
the containernetworking-plugins package.  As such, it is also
required that the --cni-bin-dir option now be passed to kubelet.

Depends-on: https://review.opendev.org/680408
Closes-Bug: #1840391
Change-Id: I9a0cbd63012d9a748826f24ae935823693048044
Signed-off-by: Steven Webster <steven.webster@windriver.com>