Update boot device query for initial install
Currently, the initial boot device for the first controller is
determined by checking the anaconda install log. This check is
problematic as the log may use a device name like /dev/sda that could
potentially change. For example, a system that has booted with the USB
enumerated as sda would need to install to sdb, but once the USB is
removed, this disk is now sda, causing a mismatch with the logs.

This update uses the /boot mountpoint instead, determining the disk on
which it is mounted and using its persistent name.

Change-Id: Ifbefc005ae1f422c1794f807460d02a02ce058db
Closes-Bug: 1843893
Signed-off-by: Don Penney <don.penney@windriver.com>