Align SR-IOV CNI templates with master
Currently, StarlingX uses a version of the SR-IOV CNI and device
plugin container images that are based on a certain commit reference.
This is done to ensure reliable and predicable behaviour until the
images can be locked down on a stable release version.

This commit aligns the ansible templates with the latest version
of these plugins, as they are being updated to pull in various
bug fixes and features.

Change-Id: If5330f794ded901cd88aa9cf2d6e13cbfbb5f062
Depends-on: https://review.opendev.org/685498
Partial-Bug: 1835020
Signed-off-by: Steven Webster <steven.webster@windriver.com>