B&R: Recover Ceph data on standard with storage nodes
This commit skips setting storage nodes to locked, offline and
skip running the recover-ceph-data role for standard deployments
that have storage nodes.

Since we are not reinstalling storage nodes (we are reinstalling
controller-0) we have up to date ceph-mon data on storage-0 and
there is no need to scan OSDs. And, since storage cluster is kept
up during the restore procedure, after controller-0 initial unlock,
ceph-mon of controller-0 connects to ceph-mon on storage-0 and
gets the up to date from there.

Change-Id: If3a1696031b157c7d740fd828ea3a56bcd8b60df
Story: 2004761
Task: 36594
Depends-On: Iabc1237410770eb12324519cfab3e6febf381e6a
Depends-On: Ifebff109f90d3091f2aa0e8958215c8c4af344ae
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>