Backup & Restore: Failed to unlock controller-0 after platform restore
This issue happened only when https is enabled.

Controller-0 failed to unlock because lighttpd.service was not up and
that caused "helm repo add" command to fail when applying controller
manifest. Because https is enabled, lighttpd service needs to access
server-cert.pem certificate to start. Even though this certification
is backed up but it is not restored during platform restore.

Change-Id: I0d7915bc95064974675614be8eb4b15cb091f684
Closes-Bug: 1845504
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>