Bootstrap fails due to incorrect minimum range calculations
The fix for https://bugs.launchpad.net/starlingx/+bug/1845215 introduced
an issue with the default value for the management_subnet. The minimum
network range is calculated as the minimum subnet size -2. However for
many networks we use the default start address as subnet start + 2.

For now we will update the default management_subnet to be /24 to avoid
this issue. Another commit will be coming to change the default start
addresses to subnet start + 1.

Change-Id: Ic0e93b78288f1e0ceceb068d0b5cae1f1dc75e89
Partial-Bug: 1848376
Signed-off-by: David Sullivan <david.sullivan@windriver.com>