Bootstrap fails due to incorrect minimum range calculations
The fix for https://bugs.launchpad.net/starlingx/+bug/1845215 exposed
an issue with the default value for the management_subnet. The minimum
network range is calculated as the minimum subnet size -2. However for
many networks we use the default start address as subnet start + 2.

Here we change the default start addresses to subnet start + 1.

Closes-Bug: 1848376
Change-Id: I81aab6e8d60de08dbf27db5b8ff8533104cb4738
Signed-off-by: David Sullivan <david.sullivan@windriver.com>