B&R: Fix failed to unlock controller-0 after platform restore
Issue is caused by lighttpd not starting on https configurations.
This happens because restore doesn't copy server-cert.pem from the
backup. This was fixed by https://review.opendev.org/#/c/685390/,
problem is it got reverted in https://review.opendev.org/#/c/686057/
in a refactoring.

This commit adds back the fix.

Change-Id: I1e8a15bc95064974675614be8eb4b15cb091f685
Closes-Bug: 1849379
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>