Fix for missing mount when kubeadm init invoked
When kubeadm reset is invoked, the /var/lib/kubelet
mount is removed.

The "when" conditional in the ansible environment did not seem
to detect that the mount command was needed.

Updated the command to be a compound bash command to ensure
it works in local and remote ansible environments.

Change-Id: Ic81d180df9691161e0b75ab7cdeaf2f639b47728
Fixes-Bug: 1849710
Related-Bug: 1847147
Related-Bug: 1838692
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>