Update the image generation logic from playbook
This commit removes the logic that updates image reference with
docker registry if the registry is not in one of the override-able
public registries.

Since we should support to download additional images via the
parameter additional_local_registry_images from a registry that
is not in one of the public registries.

Change-Id: I2e47e04de46c22c6deb7de6363e9262ac5eb047c
Closes-Bug: 1846799
Signed-off-by: Angie Wang <angie.wang@windriver.com>