Set proper permission for /opt/platform/sysinv
Based on reviews https://review.opendev.org/#/c/692439/2 and
https://review.opendev.org/#/c/691714 the owner of the
/opt/platform/sysinv folder was the "sysinv" user.

However, in https://review.opendev.org/#/c/694774/1 the
owner what changed to root:root. This cause the sysinv
conductor processs to not be able to write to this folder
as it does not have proper permissions.

The solution is to change the folder owner back to sysinv.

Change-Id: I74962ea69dfd6e9ea24e4fa51fc4e8bb9b6325ac
Closes-bug: 1853642
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>