Fix subcloud bootstrap failure
Subcloud bootstrap failed caused by an error "no attribute
'username'" generated due to using an aws erc registry for
the system images.

The aws ecr user credentials are returned from a python script,
and the return value is a dictionary which is written to the
stdout using print().  The stdout contains a preceding, and
trailing blank line when Ansible runs a playbook on a remote
host. As such, the return value was not treated as a dictionary
and attempts to reference 'username' field in the user
credentials dictionary resulted in failure.

This update strips both leading and trailing \r\n from the
user credentials output.

Tested running a playbook locally and remotely using an aws
erc registry for system images.

Closes-Bug: 1855129

Change-Id: I6cfe1163586d60ecbd0d02ef26a3f984485cc916
Signed-off-by: Tao Liu <tao.liu@windriver.com>