B&R: Add backup & restore for Helm overrides
Right now, if the user wants to restore only the stx-openstack,
the Helm overrides will be regenerated at every restore procedure,
so the Helm overrides data in the sysinv database will always change.
On the other hand, the stx-openstack database will be unchanged
at every restore and it contains some of the initial
overrides data (e.g credentials).
The issue here is that the sysinv database will be inconsistent
with the stx-openstack database.

We can fix this by adding the B&R of the Helm overrides
in the stx-openstack B&R. This will be done by
creating a database dump file which will
contain the Helm overrides data.

Change-Id: I3323a8500ace5e25be57dd5bd4bff51df609e968
Story: 2006770
Task: 37781
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>