Patch tiller deployment to provide environment validation
There appears to be a race condition between when kubelet sees a pod and
when kubelet sees a service. Due to this race, required environment
variable are missing to allow tiller to function properly.

See the comment at
https://github.com/kubernetes/kubernetes/blob/v1.18.1/pkg/kubelet/kubelet_pods.go#L566

This change patches the tiller deployment to make sure the four classes
of environment variables are present prior to starting tiller. If any
class of variables are not present in the environment, then exit. This
will recreate the pod and will populate the correct environment for
tiller to function.

Since the upgrade to v1.18.1, this has been seen in simplex and duplex
controller configurations.

This will cover patching during initial provisioning via ansible and
will be reverted once StarlingX moves to helm v3.

Change-Id: I78e43459fedab611a67b8d9b6b3121b78ef048a6
Partial-Bug: #1856078
Signed-off-by: Robert Church <robert.church@windriver.com>