Resize DRBD resources when doing a restore
In cases where we do a backup of a system that has non-default
sizes for drbd-backed partitions, the restore fails when first
unlocking controller-0.

The normal resize procedure requires all controller nodes to
be unlocked and available because the puppet manifest does
not support resizing at unlock.

To prevent the issue from occuring, as part of the restore
procedure, we should resize the partitions on controller-0
with the proper sizes found in sysinv. Controller-1 will
automatically create the partitions with the proper sizes
from the very start, so it will not need any resizes.

Change-Id: Ia73452ce721514d393b486a659730d0ca7c0d7e5
Closes-bug: 1854169
Depends-on: https://review.opendev.org/#/c/699990
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>