Add an install playbook
This update adds an install playbook that leverages
Redfish Virtual Media Controller(rvmc).

The playbook creates the rvmc resource which installs
a host via a Redfish supported Board Management Controller.
The playbook then monitors the boot progress, changes the
initial password once the host is installed/booted,
and awaits the system to be ready for the bootstrapping.

Story: 2006980
Task: 37715

Depends-On: https://review.opendev.org/#/c/702786/
Change-Id: Ib8a52fc1fe582fd01400d66653a31d8c48ee7792
Signed-off-by: Tao Liu <tao.liu@windriver.com>