Netapp: Fix handling of IPv6 addresses
Using bash process subtitution to pass the file parameter
to the "create backend" command doesn't work as the bash
variable expansion breaks the IPv6 format, resulting in
an invalid JSON file.

In order to fix this, we create a temporary file where we
store the desired information. This file is then cleaned
(along with the whole staging folder) once the playbook
finishes exection.

Also improved handling of IPv6 by the install command by
supporing extra parameters for the "tridentctl install"
command. In order to force IPv6 compatibility, you need to
add to the "trident_install_extra_params" variable the
"--use-ipv6" option

Change-Id: I5ecd98307d40cfdfe277d72acd7b095ff0b66aad
Closes-bug: 1868103
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>