Fix Netapp port conflict
By default, the Trident Netapp service opens port 8443 for
HTTPS REST api usage. This conflicts with the port the
Horizon dashboard uses on an HTTPS enabled setup (the port
is also 8443).

In order to fix this, we change the default port from 8443
to 8678, but also make it configurable through ansible
overrides.

The Trident service also opens port 8001 for metrics usage.
While that doesn't currently conflict with any other service
on the system, I also made that configurable through
ansible overrides, in case such a conflict appears in the
future.

Change-Id: I08db939acac6082f82b9e12e932d8289c7cecdeb
Closes-bug: 1868382
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>