Backup & restore: Restore license files
STX offers support for installing license files through the
"system license-install" command.

While, these licenses are not enforced, they are part of the
backups created, but they are not restored when doing a full
backup & restore.

Since license is optional, it is not expected to always be
present in the backup archive, so we only restore it if it
is present in the archive.

Change-Id: Ibd4cdcb53d1d55409d947c1f3af45659ed21a7ae
Closes-bug: 1871034
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>