Enable applying applications after bootstrap
This commit adds the ability to specify applications to be applied
directly after bootstrap, before controller-0 have been unlocked.
This is needed for cert manager.

Currently, nginx and cert manager will be applied by default, with
no overrides. The user can optionally specify overrides if they wish

NOTE: This aligns with long term direction for platform applications
to:
- move away from the existing platform application framework in sysinv
  due to wanting to decouple application behaviour from sysinv code
  in order to support such things as independent upgrades of these
  platform applications.
- support auto-upload/apply of platform applications in either:
     a) bootstrap playbook, if app required for supporting bootstrap
        functions, or
     b) a post-bootstrap deployment-type playbook.
In the case of cert-manager, in near future, it will be required at
bootstrap to support initial configuration around generating
certificates for kubernetes and https connections.

Story: 2007360
Task: 39471

Change-Id: I91ee31c7c2d35c2a101b156ef8633fc69139938d
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>