Increase wait time for certificate during subcloud bootstrap
Currently during subcloud ansible bootstrap, it waits up to 15s for
certificate secret to be ready after the yaml file applies. For some
slow hosts (VBox for example) 15s appears not long enough so the
extracted certificate is partial, which in turn fails haproxy.

This commit updates to use the better "kubectl wait" mechanism to wait
for the certificate to be ready, with a timeout of 30s.

Change-Id: Ibd8cab9339c6d532353b45b49cc4d141f0cf5ace
Closes-Bug: 1876099
Signed-off-by: Andy Ning <andy.ning@windriver.com>