Provide an update strategy for Tiller deployment
In the case of a simplex controller configuration the current patching
strategy for the Tiller environment will fail as the tiller ports will
be in use when the new deployment is attempted to be applied. The
resulting tiller pod will be stuck in a Pending state.

This will be observed if the node becomes ready after 'helm init'
installs the initial deployment and before the deployment is patched for
environment checks.

The deployment strategy provided by 'helm init' is unspecified. This
change will allow one additional pod (current + new) and one unavailable
pod (current) during an update. The maxUnavailable setting allows the
tiller pod to be deleted which will release its ports, thus allowing the
patch deployment to spin up an new pod to a Running state.

Change-Id: I83c43c52a77bce9f085bfb6c6a2c4171f2ba8f97
Partial-Bug: #1876396
Signed-off-by: Robert Church <robert.church@windriver.com>