Add kube-apiserver port to calico failsafe rules
An invalid GlobalNetworkPolicy or NetworkPolicy may prevent
calico-node from communicating with the kube-apiserver.
Once the communication is broken, calico-node is no longer
able to update the policies since it cannot communicate to
read the updated policies.  It can also prevent the pod
from starting since the policies will prevent it from
reading the configuration.

To ensure that this scenario does not happen, the kube-apiserver
port is being added to the failsafe rules to ensure communication
is always possible, regardless of the network policy configuration.

Change-Id: I1b065a74e7ad0ba9b1fdba4b63136b97efbe98ce
Closes-Bug: 1877166
Related-Bug: 1877383
Signed-off-by: Matt Peters <matt.peters@windriver.com>