Fix B&R folder permission
When /opt/backups is mounted it is given 0755 perms.
User postgres has access to the files in the dir.
During unlock this is changed to 0750.

There is a commit which changes the perms to 0750 during bootstrap.
User postgres loses access to the dir.
https://review.opendev.org/#/c/724270/

puppet.log: Executing 'chmod 0750 /opt/backups'

The fix is to change permissions during restore to 0755 as it was
before.
During unlock this will be changed to 0750.

Closes-Bug: 1880628
Change-Id: I5823c1fe7eb36bb80e7a735749a278aa501ec8ca
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>