Update containerd registry.local configuration
As part of bootstrap, k8s.gcr.io/pause:3.2 is pulled via crictl from
registry.local with explicitly provided credentials. If this image is
manually removed or removed due to garbage collection, containerd is
unable to pull it from registry.local.

Lookup the registry credentials so that they can be applied to the
registry.local auth configuration in containerd's config.toml. This will
allow containerd pull access when needed.

Change-Id: I5095abbe44c4e9bab36726a336654284482e44b4
Depends-On: https://review.opendev.org/#/c/733941/
Closes-Bug: #1881353
Signed-off-by: Robert Church <robert.church@windriver.com>