Update containerd registry.local configuration
As part of bootstrap, k8s.gcr.io/pause:3.2 is pulled via crictl from
registry.local with explicitly provided credentials. If this image is
manually removed or removed due to garbage collection, containerd is
unable to pull it from registry.local.

Lookup the registry credentials so that they can be applied to the
registry.local auth configuration in containerd's config.toml. This will
allow containerd pull access when needed.

Change-Id: Ie29a797a09879d7dff28356a2335980ab6c49bed
Depends-On: https://review.opendev.org/#/c/732724/
Closes-Bug: #1881353
Signed-off-by: Robert Church <robert.church@windriver.com>