Add test for cert-manager resource readiness
In some cases, DC installations were trying to create
cert-manager resources before app was ready. This adds a
check to ensure cert-manager is applied and a test issuer
can be created before moving on to next phase.

While Armada claims all pods are ready (the readiness check
for the pods transitions to Ready state), there are some aspects
of the app that aren't ready to use. An issue was raised in
upstream community, and it was confirmed that this type of
indirect check was the only way to confirm readiness.
See GitHub Issue here: https://github.com/jetstack/cert-manager/issues/3045

Testing was performed with installs on AIO, standard & DC
systems. B&R testing was also performed.

Closes-bug: 1881606
Closes-bug: 1880574

Change-Id: Iaf84b43bb4ce20476a9bc66b4ad7ced21753e0ff
Signed-off-by: Sabeel Ansari <Sabeel.Ansari@windriver.com>