Create rvmc in namespace with registry, hostnetwork and psp
This updates the rvmc from 'default' to 'rvmc' namespaces.
The docker registry key is copied to namespace in order to
allow the pod to pull the image.  On an upgrade, the rvmc
may no longer be in the docker image cache, and the
ImagePullSecret is required to pull it.

Enable RoleBinding in rvmc for pod security plugin.

hostNetwork is also explicitly set to allow rvmc pod
access to bmc_address.

Tests Performed:
- verify rvmc can pull from docker registry
- verify namespace is created if not already created
- verify default-registry key is copied to rvmc namespace
- verify job and pods are deleted if it exists
- apply install.yml and verify access to bmc_address
- verify with and without pod security policy plugin
    - kubectl apply -f /usr/share/ansible/stx-ansible/ \
        playbooks/roles/bootstrap/bringup-essential-services/\
        files/psp-policies.yaml
    - system service-parameter-add kubernetes kube_apiserver \
        admission_plugins=PodSecurityPolicy
    - system service-parameter-apply kubernetes

Change-Id: I43005542d15600926c7074c08773448f9f835b6c
Story: 2007403
Task: 40150
Signed-off-by: John Kung <john.kung@windriver.com>