Use controller-0 ip for calico-node can-reach dest
Because cluster floating IP host can change, it should not be used for
calico node address, and doing so has been observed to cause an error
when calico-node is rebooted on the active controller and then a swact
is executed, causing BGP peering to be lost.

This commit switches to using controller-0 cluster host address for
route selection, in order to ensure a consistent route selection is
used.

Closes-Bug: 1885582
Change-Id: I56c5ddf657eb557b83ce0fd3ce7beb71011d6266
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>