Improve ansible error message if bootstrap manifest fails to apply
This commit improves the error message for puppet manifest apply
failure to notify the host should be re-installed before attempting to
bootstrap it again.

Test:
  This message will show as the task "Fail if puppet manifest apply
script returns an error" if the task "Applying puppet bootstrap manifest"
failed with the error message "...Please re-install host {address} before
attempting to bootstrap it again...". If the previous task is
succeeded, the error message will not show.

Change-Id: Ic11ff92e6e02c0b4c95446c7381ea4123a9eec16
Close-bug: 1886856
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>