B&R: Backup user images from local registry.
The procedure is optional and is activated by setting
backup_user_local_registry=true for the backup playbook.

An user image is one that is not a platform or an app image.
Information about platform images is kept in the playbooks.
Information about apps images is retrieved through sysinv-utils.
In fact apps images are filtered out in sysinv-utils by passing
`--no-apps` arg.

Save user images to local filesystem using docker pull.
Check if there is enough disk space for the archive to be saved.
Export to a tar archive using docker save.

For remote play the copy of user images generated archive is not
implemented yet.

Partial-Bug: 1886152
Depends-On: Ic3f7a82171dccaa15a76b2578135163dc78201f3
Depends-On: I0852702be7d47cd5173a7815f33812e75332c63a
Change-Id: I4644784ea4164134f163d218e69dc4ceb148985a
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>