Add required applications to applications list automatically
We used to have a required applications list in ansible playbook. This
is checked against the user specified overrides file to ensure entries
for the required applications are present. This creates an issue when
the required applications versions are changed. If the required
applications list versions are not changed, bootstrap fails. This
commit changes the behavior of the setup. There is no longer a
required applications list with version numbers; the required
applications versions will be automatically detected by searching for
nginx and cert manager tarballs on the system. If the user overrides
file does not contain entries for these applications, default entries
with no overrides will be generated to keep things consistent with the
old setup.

Closes-bug: 1886742

Change-Id: I66e6a1c1934df64988212a960bccba8efcc62bd6
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>