Add missing quotation mark for admin endpoint certificate CN
In the ansible admin endpoint certificate setup template
(setup-sc-adminep-certs.yaml.j2), controller_floating_address as the
commonName is not quoted. This causes ansible play to fail if the
floating IP address is IPv6 and has a trailing ":" at the end.

This change added the missing quotation mark around the
controller_floating_address.

This change also added missing quotation marks for
the some variables in armada-overrides.yaml.j2.

Change-Id: Ib46d76e19b13d2144e39d4598a7ccec2808988c6
Closes-Bug: 1891904
Signed-off-by: Andy Ning <andy.ning@windriver.com>