B&R: Fix restore with upgraded platform-integ-apps
In case multiple platform-integ-apps are found in the database, the
restore code modifies for each of them the status to uploaded. The old
ones are in inactive status, but this is modified to uploaded.
This is not the intended behavior.

This crashes the periodic thread responsible for auto-appying the app.
It uses a query that expects only one entry in the database to match.
More entries match, thus the query throws an exception.
Note: The query already filters out the inactive apps.

The fix is to change the restore code to modify the status of
platform-integ-apps to uploaded but only for the one app which should
be active(the app status is different than inactive).

Closes-Bug: 1893309
Change-Id: I1492d417c2cdc9ac85858bef28663852dbe2b2fc
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>