Rework restore functionality for 4 or more nodes
This commit reworks the tasks in the bringup-essential-services role to
handle cases where a system with 4 or more nodes would be unable to
bring up the required services.

Testing performed:
- Backup and restore on AIO-DX+ (5 nodes)
- Backup and restore on AIO-SX (Single node)
- Fresh install (AIO-SX)
- Bootstrap replay (AIO-SX)

The problem is caused by the kube-controller-manager placing the cluster
zone in the PartialDisruption state which causes pods to never be
evicted from downed nodes. This state can be reached when a single node
is up (ie. controller-0 during restore) and 3 or more nodes are down. In
this state, essential service pods like calico-kube-controllers, armada,
coredns etc, might be stuck on one of the nodes that are down and will
not be evicted to be recreated on controller-0. This is most commonly an
issue for pods running as a deployment.

The fix provided here is to scale the required deployments down to 0 and
then scale them back to their required values. This forces the pods to
be rescheduled onto the active node and allows the services to become
available.

This fix also reworks how we are checking the availability of services
running as deployments. The status of the deployment is now checked
rather than just the presence of pods.

Coredns was previously being checked via grepping for a running pod, but
this could give a false-positive in the PartialDisruption state, because
the active node during restore would see coredns as running on one of
the downed nodes. We now check the status of the deployment.

Other possible approaches would be to alter the unhealthy-zone-threshold
value for kube-controller-manager or by setting controller-0 as part of
a different k8s zone. Both of these approaches would cause every
deployment pod on the downed nodes to be evicted to the active node, and
this behaviour is probably unwanted for heavily utilized systems.

Testing also revealed an issue with the task "Get wait tasks results"
where restoring a deployment with many nodes would time out before the
asynchronous tasks that check for k8s components completed. Reworked the
task to wait for a longer amount of time based on the number of nodes in
the system when performing a restore.

Closes-Bug: 1893149

Change-Id: I641125ab1c32dcc55d46009efd4654b6e388d621
Signed-off-by: Cole Walker <cole.walker@windriver.com>