Back up local registry images for platform-integ-apps on DC central
Rbd-provisioner images are not backed up on a DC central w/o Ceph even
when they are added to additional_local_registry_images. This means
that on restore of DC central they are no longer present in
registry.local. This causes subclouds with Ceph enabled to fail restore
(or apply) as they pull rbd-provisioner from central.

This commit fixes this.

Change-Id: Iabc10f694db6b53881646d1a3ee61b2a6f833d1a
Closes-Bug: 1896272
Depends-On: I12310f694db6b53881646d1a3ee61b2a6f833d1b
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>