Change restore procedure
Do not force platform-integ-apps to uploaded state.
The implication of this change is that if the backup was taken when
platform-integ-apps was not in applied state, then at the end of the
restore procedure a manual apply of platform-integ-apps is needed.
Otherwise the restore procedure takes care of it.

As one of the final steps of the restore playbook place the system in
the restore state.

This commit is the last one in a series of fixes for the automatic
applies of apps and restore procedure. This commit keeps track of the
other commits.
CLI commands to control the restore state were provided.
Changes to the periodic thread to support the new restore procedure were done.

Depends-On: I7b7fab99d457056032dbbd612363cd5036736cda
Depends-On: I44fc4aaa528e372a84115714f271b4f5e063f86e
Depends-On: I64307f75d3e693dfdb0f15828743d196aaa7ae04
Depends-On: I4d5b552c4e73f0918bbd46206dd2f8deb5d6039a

Closes-Bug: 1887648
Change-Id: I8b978a341a388ffcdd154d39dc52b93e09658ef0
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>