Fix ceph ansible restore failure when backup is taken on controller-1
Ansible playbook starts ceph through '/etc/init.d/ceph start'.
This script has two detection mechanisms for osds:
 1. it looks in the ceph.conf for \[osd\.[0-9]*\] sections, grabs
    the osd id, mounts the corresponding devices and starts ceph-osd
    daemons;
 2. it lists all folders in /var/lib/ceph/osd/*, grabs the osd id
    and starts the corresponding daemon.

When backup is taken on controller-1 it contains /etc/ceph/ceph.conf
with osds of this node. Restore is always done from controller-0
where ansible extracts ceph.conf from the backup to
/etc/ceph/ceph.conf. This leads to osds from controller-1 trying to
start on controller-0 and to ansible failure.

To fix it we remove the osds configuration from ceph.conf. This works
as we have code in the restore playbook that scans the disks for osds
and mount them in /var/lib/ceph/osd/* allowing 'etc/init.d/ceph start'
to initialize the correct ceph-osd daemons for controller-0.

Closes-Bug: 1899444
Change-Id: I10672613fc26807e0cf28ac8df5a08287d80c17a