Enable multus logging in configmap
Provides the required parameters to enable multus debug logging to a file
on the host machine where the multus-cni container runs. Logs are saved
to /var/log/multus.log

Updates the name of the configmap with a version indicator to ensure
that the multus daemonset is restarted when the new configmap is
applied.

A related review has been submitted for the logrotate configuration
to manage the size and lifetime of the log file.

Closes-Bug: 1900065

Depends-On: https://review.opendev.org/#/c/758871/
Change-Id: Id712637955a7421e1ecad55fb2c5747ed41b5c45
Signed-off-by: Cole Walker <cole.walker@windriver.com>