Wait for drbdadm prior to resize2fs filesystem
After ansible bootstrap, on the host-unlock,
"Failed to make drbd-platform primary" is observed in daemon.log resulting
in error bringing up drbd filesystems.
This is reproducible on SystemController.

Update to perform, drbd-resize on specific resource. After issuing
the drbdadm resize, a pause is also required prior to performing the
resize2fs operation.

There does not appear to be much observability into drbdadm resize
at /proc/drbd or drbd-overview, so a sleep was introduced. The pause needed
to be at least 1 second as per observations in virtual and hardware labs,
AIO and Standard controllers.

Change-Id: I7b9907092350bf677df30b1d7c54915711791fb2
Closes-Bug: 1902232
Signed-off-by: John Kung <john.kung@windriver.com>