Check local pods only to prevent B&R timeout
This fix reworks a kubectl wait command to only wait for ready pods on
the local controller. This fixes an issue where restoring a cluster with
many nodes can timeout during "Start wait for armada,
calico-kube-controllers & coredns deployments to reach Available state".
This timeout was happening because the kubectl command was waiting for
pods on downed nodes to become ready for 30 seconds per pod. In cases
where there were 5 or more unreachable pods, the async timeout value of
120 seconds would be reached and the kubectl wait commands would be
terminated before completion. This prevented subsequent Ansible tasks
from completing and resulted in a parsing error during "Fail if any of
the Kubernetes component, Networking or Armada pods are not ready by
this time"

The fix here adds --field-selector spec.nodeName=$(hostname) to the
kubectl wait command and causes only pods on the running
controller to be checked. This also ensures that the asyncronous tasks
will always only take 30 seconds, regardless of the number of nodes in
the cluser.

Removed the behaviour where the wait time would scale up based on the
number of nodes, as the return time of the async tasks is now always a
fixed 30 seconds.

Closes-Bug: 1905788

Change-Id: Idb7b30891d4fd00901aaa69412ef4c59913e21f3
Signed-off-by: Cole Walker <cole.walker@windriver.com>